new: port `uprobe-set` policytest to test `uint64` argValue
Test that the printed line by the tester-progs sees the `uint64` value.

+++ contrib/tester-progs/uprobe-simple.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 int
 __attribute__((noinline))
@@ -17,9 +18,9 @@ lasagna(int c)
 
 int
 __attribute__((noinline))
-manyargs(int zero, int one, int two, int three, int four, int five, int six, int seven, int retargidx)
+manyargs(int zero, int one, int two, int three, int four, uint64_t five, int six, uint64_t seven, int retargidx)
 {
-	printf("manyargs was passed: %d,%d,%d,%d,%d,%d,%d,%d and retargidx: %d\n", zero, one, two, three, four, five, six, seven, retargidx);
+	printf("manyargs was passed: %d,%d,%d,%d,%d,%ld,%d,%ld and retargidx: %d\n", zero, one, two, three, four, five, six, seven, retargidx);
 
 	switch (retargidx) {
 		case 0:
@@ -33,11 +34,11 @@ manyargs(int zero, int one, int two, int three, int four, int five, int six, int
 		case 4:
 			return four;
 		case 5:
-			return five;
+			return (int)five;
 		case 6:
 			return six;
 		case 7:
-			return seven;
+			return (int)seven;
 		default:
 			fprintf(stderr, "regargidx must be between 0 and 7");
 			return -1;
